Move the unit tests to check_leo.py
